pyproject.toml
[project]
name = "timeline"
version = "0.1.0"
description = ""
authors = [
    {name = "kundziorro",email = "grzybek2413@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "transformers (>=4.48.2,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "newsapi-python (>=0.2.7,<0.3.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "ruff (>=0.9.4,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
